feat: track checkpoints and allow exporting (#39)
This implements the plan described in [this issue](https://github.com/snapshot-labs/sx-api/issues/36). 

Previously, Checkpoint required a static list of block numbers to help speed up indexing a list of contracts. This changes moves that block tracking logic to the database and exposes methods on the `Checkpoint` class to seed and export checkpoints.

Here is an highlight of relevant changes in this PR:
- Created a `CheckpointsStore` class and moved all MySQL interactions related to `_*` core tables into that class. This makes things neater and will make it easier to update and test in the future.
- Implemented `Checkpoint.seedCheckpoint()` and `Checkpoint.exportCheckpoints()` to respective update and export list of blocks being tracked.
- Updated Checkpoints block indexing logic to use the new checkpointsStore to find the next blocks, and track relevant blocks.
- Moved Checkpoint Type definitions to a `types.ts` file to reduce less noise and prevent cyclic dependencies when being used by utility functions. It's also easier to export all core types this way.
- Exposed a `/_checkpoint` route to preview checkpoint blocks (not entirely sure if we want to leave this at the end, but was is good for monitoring while developing).


This was already getting large, so I figured this will be a great place to open a PR, everything is tested manually, but this would benefit from automated tests, and I'll be doing that in future PRs. I'll definitely try and create smaller PR in the future 😅